+++ scripts/recon/gradient_descent.py
@@ -13,7 +13,6 @@ import os
 import numpy as np
 import time
 import pathlib as plib
-from datetime import datetime
 import matplotlib.pyplot as plt
 from lensless.io import load_data
 from lensless import (
@@ -28,7 +27,6 @@ from lensless import (
 def gradient_descent(
     config,
 ):
-
     psf, data = load_data(
         psf_fp=to_absolute_path(config.input.psf),
         data_fp=to_absolute_path(config.input.data),